requirements.txt
fastapi
redis
celery
uvicorn
python-dotenv
pydantic-settings
bittensor
httpx
